Final version
This version creates a span tag that contains an img tag. This is so I can set the img opacity separately from the animatated span opacity. The previous version used object to include the svg, but that created huge performance issues, probably related to memory leaks in some extensions like LastPass. This solves the issue for good.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>